requirements.txt
PyMuPDF
google-genai
python-docx
Pillow
python-dotenv
